% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{data.default}
\alias{data.default}
\title{export}
\usage{
\method{data}{default}(..., list = character(), package = NULL,
  lib.loc = NULL, verbose = getOption("verbose"), envir = .GlobalEnv)
}
\description{
export
}

