% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{distname.SpThin}
\alias{distname}
\alias{distname.SpThin}
\title{Distance metric used to thin data.}
\usage{
\method{distname}{SpThin}(x)

distname(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{...}{not used.}
}
\value{
\code{character} name of distance metric used to thin records.
}
\description{
This function returns the name of the distance metric used to thin datasets contained in a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
# make thinned dataset using simulated data
result <- spThin(
		runif(100, -5, -5),
		runif(100, -5, -5),
		dist=5,
		method='heuristic',
		1,
)

# show distance name of metric
distname(result)

export
}

