% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpPartial.R, R/generics.R
\name{[[.SpPartial}
\alias{[[.SpPartial}
\alias{extract}
\title{Extract samples from \code{SpRarefy} and \code{SpThin} object}
\usage{
\method{[[}{SpPartial}(x, r = 1, ...)
}
\arguments{
\item{x}{\code{SpThin} or \code{SpRarefy} object.}

\item{r}{\code{numeric} n'th replicate to extract samples from. Zero can be used to extract samples using the best replicate.}

\item{...}{not used.}
}
\value{
\code{SpatialPoints} or  \code{SpatialPointsDataFrame} object
}
\description{
This function extracts samples from a \code{SpRarefy} or a \code{SpThin} object.
}
\examples{
# make simulated dataset
testDF <- data.frame(
		x=runif(100, -5, -5),
		y=runif(100, -5, -5)
)

# make thinned dataset
result1 <- spThin(
		testDF,
		"x",
		"y",
		dist=5,
		method='heuristic',
		nrep=10
)

# extract first replicate dataset in result1
result1[[1]]

# make rarefied dataset
result2 <- spRarefy(
		testDF,
		"x",
		"y",
		grid=5,
		nrep=10
)

# extract first replicate dataset in result2
result2[[1]]
}
\seealso{
\code{\link{SpThin}}.
}

