% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpPartial.R, R/generics.R
\name{fulldata.SpPartial}
\alias{fulldata}
\alias{fulldata.SpPartial}
\title{Full dataset contained in a \code{SpThin} and \code{SpRarefy} and  object}
\usage{
\method{fulldata}{SpPartial}(x, ...)

fulldata(x, ...)
}
\arguments{
\item{x}{\code{SpRarefy} or a \code{SpThin} object.}

\item{...}{not used.}
}
\value{
\code{SpatialPoints} or \code{SpatialPointsDataFrame} object.
}
\description{
This function returns the initial dataset subject to spatial processing in a \code{SpRarefy} or a \code{SpThin} object.
}
\examples{
# make simulated dataset
testDF <- data.frame(
		x=runif(100, -5, -5),
		y=runif(100, -5, -5)
)

# make thinned dataset
result1 <- spThin(
		testDF,
		"x",
		"y",
		mindist=5,
		method='heuristic',
		1
)

# show first rows of full dataset in result1
head(fulldata(result1))

# make rarefied dataset
result2 <- spRarefy(
		testDF,
		"x",
		"y",
		grid=5,
		1
)

# show first rows of full dataset in result2
head(fulldata(result2))
}
\seealso{
\code{\link{SpThin}}.
}

