% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.grid.R
\name{make.grid}
\alias{make.grid}
\title{Make a \code{RasterLayer} based on \code{extent} and a cell size}
\usage{
make.grid(x, cellsize, crs)
}
\arguments{
\item{x}{\code{Extent} object.}

\item{cellsize}{\code{numeric} cell size for each cell.}

\item{crs}{\code{CRS} object representing coordinate system. Defaults to NA.}
}
\value{
\code{RasterLayer}.
}
\description{
This function generates a \code{RasterLayer} using an \code{extent} object and a cell size.
}
\seealso{
\code{\link[raster]{raster}}, \code{\link[sp]{CRS}}, \code{\link[raster]{Extent}}.
export
}

