% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpFilter.R
\name{plot.SpFilter}
\alias{plot.SpFilter}
\title{Map for filtered datasets}
\usage{
\method{plot}{SpFilter}(x, i = 1, ...)
}
\arguments{
\item{x}{\code{SpFilter} object.}

\item{i}{"numeric" replicate dataset to plot.}

\item{...}{not used.}
}
\description{
This function produces a map showing the distribution of occurrence records and the selected records.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spFilter(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     200000,
		10
	)

# show map
plot(result)
}
\seealso{
\code{\link{spFilter}}.
}

