% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpRarefy.R
\name{plot.SpRarefy}
\alias{plot.SpRarefy}
\title{Map for rarefied datasets}
\usage{
\method{plot}{SpRarefy}(x, i = 1, ...)
}
\arguments{
\item{x}{\code{SpRarefy} object.}

\item{i}{"numeric" replicate dataset to plot.}

\item{...}{not used.}
}
\description{
This function produces a map showing the distribution of occurrence records and the selected records.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spRarefy(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     200000,
		10
	)

# show map
plot(result)
}
\seealso{
\code{\link{spRarefy}}.
}

