% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.spThin}
\alias{plot.spThin}
\title{Plot diagnosis for results of thin function}
\usage{
\method{plot}{spThin}(thinned, which = c(1:3), ...)
}
\arguments{
\item{thinned}{A list of data.frames returned by \code{\link{thin}} function.}

\item{which}{if a subset of the plots is required, specify a subset of the numbers 1:3.}

\item{...}{other parameters to be passed through to plotting functions.}

\item{ask}{logical; if \code{TRUE}, the user is asked before each plot, see par(ask=.).}
}
\description{
Three plots (selected by \code{which}) are currently available:
a plot of the number of repetitions versus the number of maximum records retained
at each repetition ([1] observed values; [2] log transformed) and
a histogram of the maximum records retained [3].
}
\seealso{
\code{\link{thin}}
}

