% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpPartial.R, R/generics.R
\name{samples.SpPartial}
\alias{[[.SpPartial}
\alias{samples}
\alias{samples.SpPartial}
\title{Extract samples from \code{SpRarefy} and \code{SpThin} object}
\usage{
\method{samples}{SpPartial}(x, r = 1, ...)

\method{[[}{SpPartial}(x, r = 1, ...)

samples(x, ...)
}
\arguments{
\item{x}{\code{SpThin} or \code{SpRarefy} object.}

\item{r}{\code{numeric} n'th replicate to extract samples from.}

\item{...}{not used.}
}
\value{
\code{SpatialPoints} or  \code{SpatialPointsDataFrame} object
}
\description{
This function extracts samples from a \code{SpRarefy} or a \code{SpThin} object.
}
\examples{
# make simulated dataset
testDF <- data.frame(
		x=runif(100, -5, -5),
		y=runif(100, -5, -5)
)

# make thinned dataset
result1 <- spThin(
		testDF,
		"x",
		"y",
		mindist=5,
		method='heuristic',
		nrep=10
)

# extract first replicate dataset in result1
samples(result1, 1)
result1[[1]]

# make rarefied dataset
result2 <- spRarefy(
		testDF,
		"x",
		"y",
		grid=5,
		nrep=10
)

# extract first replicate dataset in result2
samples(result2, 1)
result2[[1]]
}
\seealso{
\code{\link{SpThin}}.
}

