% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpPartial.R
\docType{methods}
\name{show,SpPartial-method}
\alias{show,SpPartial-method}
\title{Show method}
\usage{
\S4method{show}{SpPartial}(object)
}
\arguments{
\item{object}{\code{SpFilter} or \code{SpThin} object.}
}
\description{
This method sets the default show method for the \code{SpFilter} and \code{SpThin} objects.
}

