% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/filter.R, R/generics.R
\name{spFilter.numeric}
\alias{spFilter}
\alias{spFilter.SpatialPoints}
\alias{spFilter.data.frame}
\alias{spFilter.numeric}
\title{Spatially filter species occurrence data}
\usage{
\method{spFilter}{numeric}(x, y, cellsize, nrep, proj4string = CRS())

\method{spFilter}{data.frame}(x, x.col, y.col, cellsize, nrep,
  proj4string = CRS())

\method{spFilter}{SpatialPoints}(x, cellsize, nrep)

spFilter(x, ...)
}
\arguments{
\item{x}{\code{numeric} with x-coordinates, \code{data.frame}, \code{SpatialPoints}, \code{SpatialPointsDataFrame} or  object.}

\item{y}{\code{numeric} with y-coordinates.}

\item{nrep}{\code{numeric} number of replicate thinned data sets to produce.}

\item{proj4string}{\code{CRS} object with coordinate system for coordinates (not needed for \code{Spatial objects}).}

\item{lon.col}{\code{character} name of column or \code{numeric} index of column with latitude values.}

\item{lat.col}{\code{character} name of column or \code{numeric} index of column with longitude values.}

\item{dist}{\code{numeric} size of grid cells. If coordinates are in a projected CRS then units are in metres.}
}
\value{
\code{SpFilter} object.
}
\description{
A grid is overlayed on a set of observation records. A random point inside each grid cell is returned. This procedure can be repeated multiple times to yield multiple filtered datasets.
This process helps reduce the effect of biases in observation records on the predictive performance of ecological niche models.
}
\seealso{
\code{\link{SpFilter}}
}

