% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpPartial.R
\docType{class}
\name{SpPartial-class}
\alias{SpPartial}
\alias{SpPartial-class}
\title{SpPartial: An S4 class to represent spatially rarefied or thinned datasets}
\description{
This class is used to store replicates of spatially rarefied thinned datasets.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{SpatialPoints} or \code{SpatialPointsDataFrame} object with all records.}

\item{\code{samples}}{\code{list} with indices for records in each replicate.}

\item{\code{call}}{\code{call} used to generate object.}
}}
\seealso{
\code{\link{call}}, \code{\link{data}}, \code{\link{nrep}}, \code{\link{plot}},\code{\link{summary}}, \code{\link{write}}.
}

