% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/rarefy.R
\name{spRarefy}
\alias{spRarefy}
\alias{spRarefy.SpatialPoints}
\alias{spRarefy.data.frame}
\alias{spRarefy.numeric}
\title{Spatially rarefy species occurrence data}
\usage{
spRarefy(x, ...)

\method{spRarefy}{numeric}(x, y, grid, nrep, proj4string = CRS(), ...)

\method{spRarefy}{data.frame}(x, x.col, y.col, grid, nrep,
  proj4string = CRS(), ...)

\method{spRarefy}{SpatialPoints}(x, grid, nrep, ...)
}
\arguments{
\item{x}{\code{numeric} with x-coordinates, \code{data.frame}, \code{SpatialPoints}, \code{SpatialPointsDataFrame} or  object.}

\item{...}{not used.}

\item{y}{\code{numeric} with y-coordinates.}

\item{grid}{Either \code{numeric} size of grid cells, \code{RasterLayer}, or \code{SpatialPolygons}, or \code{SpatialPolygonsDataFrame} object.}

\item{nrep}{\code{numeric} number of replicate thinned data sets to produce.}

\item{proj4string}{\code{CRS} object with coordinate system for coordinates (not needed for \code{Spatial objects}).}

\item{x.col}{\code{character} name of column or \code{numeric} index of column with latitude values.}

\item{y.col}{\code{character} name of column or \code{numeric} index of column with longitude values.}
}
\value{
\code{SpRarefy} object.
}
\description{
A grid is overlayed on a set of observation records. A random point inside each grid cell is returned. This procedure can be repeated multiple times to yield multiple rarefied datasets.
This process helps reduce the effect of biases in observation records on the predictive performance of ecological niche models.
}
\examples{
# load sp & raster packages
library(sp)
library(raster)

# load data
data(Heteromys_anomalus_South_America)

# set projected coordinate system
crs <- CRS(
	 	'+proj=lcc +lat_1=9 +lat_2=3 +lat_0=6
	 	+lon_0=-66 +x_0=1000000 +y_0=1000000
		+ellps=intl +towgs84=-288,175,-376,0,0,0,0 +units=m +no_defs'
)

# project lon/lat data to X/Y
sp <- spTransform(
		SpatialPoints(
			coords=as.matrix(Heteromys_anomalus_South_America[,c("LONG", "LAT")]),
			proj4string=CRS('+proj=longlat +ellps=WGS84 +datum=WGS84 +no_def')
		),
		crs
	)

# create X/Y columns with values
Heteromys_anomalus_South_America$X <- coordinates(sp)[,1]
Heteromys_anomalus_South_America$Y <- coordinates(sp)[,2]

# make rarefied dataset using numeric vectors
result1 <- spRarefy(
		Heteromys_anomalus_South_America$X,
		Heteromys_anomalus_South_America$Y,
     200000,
		10,
		proj4string=crs
)

# make rarefied dataset using data.frame
result2 <- spRarefy(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     200000,
		10,
		proj4string=crs
)

# make rarefied dataset using SpatialPointsDataFrame
result3 <-spThin(
		SpatialPointsDataFrame(
			coords=as.matrix(Heteromys_anomalus_South_America[,c("X", "Y")]),
			data=Heteromys_anomalus_South_America,
			proj4string=crs
		),
		10
)
}
\seealso{
\code{\link{SpRarefy}}
}

