% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R
\docType{class}
\name{SpThin-class}
\alias{SpThin}
\alias{SpThin-class}
\title{SpThin: An S4 class to represent spatially thinned datasets}
\description{
This class is used to store replicates of spatially thinned datasets.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{SpatialPoints} or \code{SpatialPointsDataFrame} object with all records.}

\item{\code{samples}}{\code{list} with indices for records in each replicate.}

\item{\code{mindist}}{\code{numeric} minimum distance (m) between records.}

\item{\code{method}}{\code{character} name of method used to solve problem.}

\item{\code{call}}{\code{call} used to generate object.}
}}
\seealso{
\code{\link{call}}, \code{\link{data}}, \code{\link{samples}}, \code{\link{mindist}}, \code{\link{nrep}}, \code{\link{plot}},\code{\link{summary}}, \code{\link{write}}.
}

