% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpFilter.R
\name{summary.SpFilter}
\alias{summary.SpFilter}
\title{Summarize filtered datasets}
\usage{
\method{summary}{SpFilter}(object, ...)
}
\arguments{
\item{object}{\code{SpFilter} object.}

\item{...}{not used.}
}
\description{
\code{summary} method for \code{SpFilter} objects.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spFilter(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     200000,
		10
	)

# show map
summary(result)
}
\seealso{
\code{\link{spFilter}}.
}

