% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpPartial.R
\name{summary.SpPartial}
\alias{summary.SpPartial}
\title{Summarize filtered or thinned  datasets}
\usage{
\method{summary}{SpPartial}(x)
}
\arguments{
\item{x}{\code{SpThin} or \code{SpFilter} object.}
}
\value{
\code{list} with the (1) maximum number of records, (2) number of data frames
with maximum number of records and (3) a table with the number of data frames per number of records.
}
\description{
\code{summary} method for \code{SpFilter} or \code{SpThin} objects.
}
\seealso{
\code{\link{spFilter}}, \code{\link{spThin}}.
}

