% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpRarefy.R
\name{summary.SpRarefy}
\alias{summary.SpRarefy}
\title{Summarize rarefied datasets}
\usage{
\method{summary}{SpRarefy}(object, ...)
}
\arguments{
\item{object}{\code{SpRarefy} object.}

\item{...}{not used.}
}
\description{
\code{summary} method for \code{SpRarefy} objects.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spRarefy(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     10000,
		10,
		great.circle.distance=TRUE
	)

# show map
summary(result)
}
\seealso{
\code{\link{spRarefy}}.
}

