% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.spThin}
\alias{summary.spThin}
\title{Summary method for results of thin function}
\usage{
summary.spThin(thinned, show = TRUE)
}
\arguments{
\item{thinned}{A list of data.frames returned by \code{\link{thin}} function.}

\item{show}{logical; if \code{TRUE},the summary values are printed at the console.}
}
\value{
Returns a list with the (1) maximun number of records, (2) number of data frames
with maximun number of records and (3) a table with the number of data frames per
number of records.
}
\description{
Summarize the results of \code{thin} function.
}
\seealso{
\code{\link{thin}}
}

