% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/thin.R
\name{thin}
\alias{thin}
\alias{thin.SpatialPoints}
\alias{thin.data.frame}
\title{Spatially thin species occurrence data}
\usage{
thin(x, ...)

\method{thin}{data.frame}(x, lon.col, lat.col, mindist, nrep,
  great.circle.distance = TRUE, fast = TRUE)

\method{thin}{SpatialPoints}(x, mindist, nrep,
  great.circle.distance = x@proj4string@projargs ==
  "+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs", fast = TRUE)
}
\arguments{
\item{x}{\code{SpatialPoints}, \code{SpatialPointsDataFrame} or \code{data.frame} object.}

\item{lon.col}{\code{character} name of column or \code{numeric} index of column with latitude values.}

\item{lat.col}{\code{character} name of column or \code{numeric} index of column with longitude values.}

\item{mindist}{\code{numeric} minimum distance (m) between records.}

\item{great.circle.distance}{\code{logical} if \code{TRUE} great circle distances will be used for distance calculations, else euclidean distances will be be used.}

\item{fast}{\code{logical} if \code{TRUE} the point with the most number of close neighbours will be removed at each iteration. This results in a lower run-time with less variation between replicates. Otherwise the points will be randomly sampled for removal based on their number of close neighbours. This results in a longer run-time with more variation between replicates.}

\item{reps}{\code{numeric} number of replicate thinned data sets to produce.}
}
\value{
\code{SpThin} object.
}
\description{
A heuristic algorithm is used to subset the dataset such that all occurrence
locations are a minimum distance apart. This process helps reduce the effect
of biases in observation records on the predictive performance of ecological niche models.
}
\seealso{
\code{\link{SpThin}}
}

