% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R
\name{write.default}
\alias{write.default}
\title{export}
\usage{
\method{write}{default}(x, file = "data", ncolumns = if (is.character(x)) 1
  else 5, append = FALSE, sep = " ")
}
\description{
export
}

