% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{write.SpThin}
\alias{write}
\alias{write.SpThin}
\title{Write thinned dataset replicates to file.}
\usage{
\method{write}{SpThin}(x, coords = FALSE, dir = getwd(), base = "thin_")

write(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{coords}{\code{logical} if \code{TRUE} only coordinates of thinned data will be be saved, otherwise all columns of thinned data will be saved.}

\item{dir}{\code{character} directory to save output files in.}

\item{base}{\code{character} base name to save output files in.}
}
\description{
This function writes the spatially thinned replicates in a \code{SpThin} object to files on a computer.
}
\seealso{
\code{\link{SpThin}}.
}

