% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpFilter.R
\name{write.SpFilter}
\alias{write.SpFilter}
\title{Write filtered dataset replicates to file.}
\usage{
write.SpFilter(x, coords = FALSE, dir = getwd(), base = "thin_")
}
\arguments{
\item{x}{\code{SpFilter} object.}

\item{coords}{\code{logical} if \code{TRUE} only coordinates of thinned data will be be saved, otherwise all columns of thinned data will be saved.}

\item{dir}{\code{character} directory to save output files in.}

\item{base}{\code{character} base name to save output files in.}

\item{...}{not used.}
}
\description{
This function writes the spatially thinned replicates in a \code{SpFilter} object to files on a computer.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make filtered dataset
result <-spFilter(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     200000,
		10
	)

# save data to temporary directory
write.SpFilter(result, dir=tempdir())

# show files in temporary directory
dir(tempdir())
# remove files
unlink(dir(tempdir(), '^filter_.*.csv$'))
}
\seealso{
\code{\link{SpFilter}}.
}

