% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpRarefy.R
\name{write.SpRarefy}
\alias{write.SpRarefy}
\title{Write rarefied dataset replicates to file.}
\usage{
write.SpRarefy(x, coords = FALSE, dir = getwd(), base = "thin_")
}
\arguments{
\item{x}{\code{SpRarefy} object.}

\item{coords}{\code{logical} if \code{TRUE} only coordinates of thinned data will be be saved, otherwise all columns of thinned data will be saved.}

\item{dir}{\code{character} directory to save output files in.}

\item{base}{\code{character} base name to save output files in.}

\item{...}{not used.}
}
\description{
This function writes the spatially thinned replicates in a \code{SpRarefy} object to files on a computer.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make rarefied dataset
result <-spRarefy(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
     200000,
		10
	)

# save data to temporary directory
write.SpRarefy(result, dir=tempdir())

# show files in temporary directory
dir(tempdir())
# remove files
unlink(dir(tempdir(), '^rarefy_.*.csv$'))
}
\seealso{
\code{\link{SpRarefy}}.
}

