% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpThin.R
\name{write.SpThin}
\alias{write.SpThin}
\title{Write thinned dataset replicates to file.}
\usage{
write.SpThin(x, coords = FALSE, dir = getwd(), base = "thin_")
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{coords}{\code{logical} if \code{TRUE} only coordinates of thinned data will be be saved, otherwise all columns of thinned data will be saved.}

\item{dir}{\code{character} directory to save output files in.}

\item{base}{\code{character} base name to save output files in.}

\item{...}{not used.}
}
\description{
This function writes the spatially thinned replicates in a \code{SpThin} object to files on a computer.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spThin(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
		method='heuristic',
     dist=10000,
		nrep=10,
		great.circle.distance=TRUE
	)

# save data to temporary directory
write.SpThin(result, dir=tempdir())

# show files in temporary directory
dir(tempdir())
# remove files
unlink(dir(tempdir(), '^thin_.*.csv$'))
}
\seealso{
\code{\link{SpThin}}.
}

