% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpPartial.R, R/generics.R
\name{call.SpPartial}
\alias{call}
\alias{call.SpPartial}
\title{Function call used to generate SpThin object}
\usage{
\method{call}{SpPartial}(x)

call(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}
}
\value{
\code{call} object.
}
\description{
This function returns the function call used to generate a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
export
}

