% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpFilter.R, R/generics.R
\name{cellsize.SpFilter}
\alias{cellsize}
\alias{cellsize.SpFilter}
\title{Cell size distance used to filter dataset.}
\usage{
\method{cellsize}{SpFilter}(x, ...)

cellsize(x, ...)
}
\arguments{
\item{x}{\code{SpFilter} object.}

\item{...}{not used.}
}
\value{
\code{numeric} cell size (m) used to filter records.
}
\description{
This function returns the cell size used to filter records in a \code{SpFilter} object.
}
\examples{
# make filtered dataset using simulated data
result <- spFilter(
		runif(100, -5, -5),
		runif(100, -5, -5),
		grid=5,
		1,
)

# show cell size
cellsize(result)
}
\seealso{
\code{\link{SpFilter}}.
}

