% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpRarefy.R, R/generics.R
\name{cellsize.SpRarefy}
\alias{cellsize}
\alias{cellsize.SpRarefy}
\title{Cell size distance used to rarefy dataset.}
\usage{
\method{cellsize}{SpRarefy}(x, ...)

cellsize(x, ...)
}
\arguments{
\item{x}{\code{SpRarefy} object.}

\item{...}{not used.}
}
\value{
\code{numeric} cell size (m) used to rarefy records.
}
\description{
This function returns the cell size used to rarefy records in a \code{SpRarefy} object.
}
\examples{
# make rarefied dataset using simulated data
result <- spRarefy(
		runif(100, -5, -5),
		runif(100, -5, -5),
		grid=5,
		1,
)

# show cell size
cellsize(result)
}
\seealso{
\code{\link{SpRarefy}}.
}

