% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{data.SpThin}
\alias{data}
\alias{data.SpThin}
\title{Complete dataset contained in a SpThin object}
\usage{
\method{data}{SpThin}(x)

data(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}
}
\value{
\code{call} object.
}
\description{
This function returns the complete dataset subject to spatial thinning in a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
}

