% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Heteromys_anomalus_South_America.R
\docType{data}
\name{Heteromys_anomalus_South_America}
\alias{Heteromys_anomalus_South_America}
\title{Occurrence record locations for Heteromys anomalus}
\format{A data frame with 201 rows and 4 variables}
\description{
A dataset containing compiled occurrence record locations for
Heteromys anomalus in northern coastal South America. These records
have been examined to check for accurate species identification.
}
\details{
\itemize{
  \item \code{SPEC}: species name assigned to occurrence record
  \item \code{LAT}: decimal degree latitude value
  \item \code{LONG}: decimal degree longitude value
  \item \code{REGION} region, or island, of occurrence
}
}
\examples{
data(Heteromys_anomalus_South_America)
head(Heteromys_anomalus_South_America)
}
\keyword{datasets}

