% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Heteromys_anomalus_South_America}
\alias{Heteromys_anomalus_South_America}
\title{Occurrence record locations for Heteromys anomalus}
\format{A data frame with 201 rows and 4 variables}
\description{
A dataset containing compiled occurrence record locations for
Heteromys anomalus in northern coastal South America. These records
have been examined to check for accurate species identification.
}
\details{
\itemize{
  \item SPEC. species name assigned to occurrence record
  \item LAT. decimal degree latitude value
  \item LONG. decimal degree longitude value
  \item REGION. region, or island, of occurrence
}
}
\keyword{datasets}

