% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{mindist.SpThin}
\alias{mindist}
\alias{mindist.SpThin}
\title{Minimum distance in thinned datasets.}
\usage{
\method{mindist}{SpThin}(x)

mindist(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{...}{not used.}
}
\value{
\code{numeric} minimum distance (m) between records.
}
\description{
This function returns the minimum distances in the thinned dataset contained in a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
# make thinned dataset using simulated data
result <- spThin(
		runif(100, -5, -5),
		runif(100, -5, -5),
		dist=5,
		method='heuristic',
		1,
)

# show minimum distance
mindist(result)

export
}

