% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{mindist.SpThin}
\alias{mindist}
\alias{mindist.SpThin}
\title{Minimum distance used to thin dataset.}
\usage{
\method{mindist}{SpThin}(x)

mindist(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}
}
\value{
\code{numeric} distance (m) used to thin records.
}
\description{
This function returns the minimum distance used to thin records in a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
}

