% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{nrep.SpThin}
\alias{nrep}
\alias{nrep.SpThin}
\title{Number replicates in SpThin object}
\usage{
\method{nrep}{SpThin}(x)

nrep(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}
}
\value{
\code{integer} number of replicates.
}
\description{
This function extracts the number of replicates contained in a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
}

