% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R
\name{plot.SpThin}
\alias{plot.SpThin}
\title{Map and diagnostic plots for thinned datasets}
\usage{
\method{plot}{SpThin}(x, which = 1:4, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{which}{"numeric" values to indicate which plots to produce.}

\item{...}{not used.}
}
\description{
This function produces plots to describe \code{SpThin} datasets:
\enumerate{
	\item map showing distribution of occurrence records (hexagons) and records in best solution (red points);
	\item plot of the number of repetitions versus the number of maximum records retained;
	\item log-log version of the above plot;
	\item histogram with density curve of the maximum records retained.
}
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spThin(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
		method='heuristic',
     dist=10000,
		nrep=10,
		great.circle.distance=TRUE
	)

# show map + diagnostic plots
plot(result)
}
\seealso{
\code{\link{spThin}}.
}

