% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R
\name{plot.SpThin}
\alias{plot.SpThin}
\title{Diagnostic plots for thinned datasets}
\usage{
plot.SpThin(x, which = c(1:3), ...)
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{which}{"numeric" values to indicate which plots to produce.}
}
\description{
This function produces plots to describe the thinned datasets:
\enumerate{
	\item plot of the number of repetitions versus the number of maximum records retained;
	\item log-log version of the above plot;
	\item histogram with density curvie of the maximum records retained.
}
}
\seealso{
\code{\link{thin}}
}

