% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R, R/generics.R
\name{samples.SpThin}
\alias{[[.SpThin}
\alias{samples}
\alias{samples.SpThin}
\title{Extract samples from SpThin object}
\usage{
\method{samples}{SpThin}(x, r = 1)

\method{[[}{SpThin}(x, r = 1)

samples(x, ...)
}
\arguments{
\item{x}{\code{SpThin} object.}

\item{r}{\code{numeric} n'th replicate to extract samples from.}
}
\value{
\code{SpatialPoints} or  \code{SpatialPointsDataFrame} object
}
\description{
This function extracts samples from a \code{SpThin} object.
}
\seealso{
\code{\link{SpThin}}.
}

