% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R
\docType{class}
\name{SpThin-class}
\alias{SpThin}
\alias{SpThin-class}
\title{SpThin: An S4 class to represent spatially thinned datasets}
\description{
This class is used to store replicates of spatially thinned datasets.
}
\section{Slots}{

\describe{
\item{\code{data}}{\code{SpatialPoints} or \code{SpatialPointsDataFrame} object with all records.}

\item{\code{samples}}{\code{list} with indices for records in each replicate.}

\item{\code{method}}{\code{character} name of method used to thin the data.}

\item{\code{thindist}}{\code{numeric} distance (m) used to thin records.}

\item{\code{distname}}{\code{character} name of distance metric used to thin data. Either 'great circle distance' or 'Euclidean distance'.}

\item{\code{mindist}}{\code{numeric} minimum distances (m) between records for each replicate.}

\item{\code{call}}{\code{call} used to generate object.}
}}
\seealso{
\code{\link{call}}, \code{\link{data}}, \code{\link{mindist}}, \code{\link{nrep}}, \code{\link{plot}},\code{\link{summary}}, \code{\link{thindist}}, \code{\link{write}}.
}

