\name{spThinr-package}
\alias{spThinr-package}
\alias{spThinr}
\docType{package}
\title{
Spatially thins species occurence data
}
\description{
spThin is a set of functions that can be used to spatially thin
species occurence data. The resulting thinned data can be used in ecological
modeling, such as ecological niche modeling.
}
\details{
\tabular{ll}{
Package: \tab spThinr\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2014-05-18\cr
License: \tab GPL-3\cr
}
ToDo - An overview of how to use the package, including the most important functions
}
\author{
Matthew Aiello-Lammens

Maintainer: Matthew Aiello-Lammens <matt.lammens@gmail.com>
}
\references{
ToDo - Literature or other references for background information
}
ToDo - Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
ToDo - Optional links to other man pages, e.g. ~~
\code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
ToDo - simple examples of the most important functions ~~
}
