% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R
\name{summary.SpPartial}
\alias{summary.SpPartial}
\title{Summarize thinned datasets}
\usage{
\method{summary}{SpPartial}(object, ...)
}
\arguments{
\item{object}{\code{SpThin} object.}

\item{...}{not used.}
}
\description{
\code{summary} method for \code{SpThin} objects.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spThin(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
		method='heuristic',
     200000,
		10
	)

# show summary
summary(result)
}
\seealso{
\code{\link{spThin}}.
}

