% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SpThin.R
\name{summary.SpThin}
\alias{summary.SpThin}
\title{Summarize thinned datasets}
\usage{
\method{summary}{SpThin}(x)
}
\arguments{
\item{x}{\code{SpThin} object.}
}
\value{
\code{list} with the (1) maximum number of records, (2) number of data frames
with maximum number of records and (3) a table with the number of data frames per number of records.
}
\description{
\code{summary} method for \code{SpThin} objects.
}
\seealso{
\code{\link{SpThin}}
}

