% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SpThin.R
\name{summary.SpThin}
\alias{summary.SpThin}
\title{Summarize thinned datasets}
\usage{
summary.SpThin(object, ...)
}
\arguments{
\item{object}{\code{SpThin} object.}

\item{...}{not used.}
}
\description{
\code{summary} method for \code{SpThin} objects.
}
\examples{
# load data
data(Heteromys_anomalus_South_America)

# make thinned dataset
result <-spThin(
		Heteromys_anomalus_South_America,
		x.col = "LONG",
     y.col = "LAT",
		method='heuristic',
     dist=10000,
		nrep=10,
		great.circle.distance=TRUE
	)

# show summary
summary(result)
}
\seealso{
\code{\link{spThin}}.
}

